Avoid unconditional pyarrow dependency in dataframe.backends
This avoids importing pyarrow unconditionally in dataframe.backends,
which is reachable in some dask.array workloads (notably, those using
xarray, which use pandas.Index classes, tokenization of which requires
importing this module; however that class of users doesn't depend
on dask[dataframe] and so may not have the pyarrow dependency.

To accomplish this, we need to reorganize the root of the dask.dataframe
package a bit. I've moved some uses of pandas in dask.dataframe._compat
to dask._pandas_compat, which is outside of the dask.dataframe subpackage.

This lets users of just pandas, but not dask.dataframe, (like tokenization
registration for pandas objects) import from dask._pandas_compat, while
avoiding all of dask.dataframe.

Closes https://github.com/dask/dask/issues/12072
